Crate physical_nic

Defines a trait PhysicalNic that must be implemented by any NIC driver that wants to support language-level virtualization. This trait defines functions that can be called from a VirtualNic drop handler to return NIC resources to the OS.

Traits

PhysicalNic

This trait must be implemented by any NIC driver that wants to support language-level virtualization. It provides functions that are used to return Rx/Tx queues back to the physical NIC.